requirements.txt
requests
python-dotenv
rich
schedule
pyyaml
